Ovis_finetune/requirements.txt
# -----------------------------------------
#  核心框架与依赖
# -----------------------------------------
# ModelScope Swift 
modelscope-swift 
# 如果你是从源码安装的 swift, 这行可能不需要显式写出，
# 但明确列出可以提醒项目依赖 Swift

# PyTorch (版本需与 CUDA 兼容, Swift 可能有特定版本要求)
torch
# torchvision # 根据需要，有时需要配合 torch 安装
# torchaudio  # 根据需要，有时需要配合 torch 安装

# Hugging Face 生态系统 (Swift 依赖)
transformers
peft
accelerate
datasets

# -----------------------------------------
#  模型/分词器特定依赖 
# -----------------------------------------
sentencepiece  # 很多模型的 Tokenizer 需要
protobuf       # Transformers 或其他库可能间接需要
# timm           # 某些视觉模型可能需要

# -----------------------------------------
#  数据处理与评估
# -----------------------------------------
Pillow         # 处理图像 (PIL)
editdistance   # 计算 CER / WER
nltk           # 计算 BLEU 分数

# -----------------------------------------
#  优化库 
# -----------------------------------------
# deepspeed      # 仅当使用 --deepspeed 时需要安装，版本需与 torch/cuda 兼容

# -----------------------------------------
#  工具与日志
# -----------------------------------------
tqdm           # 显示进度条
tensorboard    # 记录和可视化训练日志 (如果 --report_to tensorboard)
# wandb          # 可选: 如果使用 Weights & Biases 进行日志记录

# -----------------------------------------
#  其他常用库
# -----------------------------------------
numpy          # 常见的数值计算库
pandas         # 可能用于数据分析或处理
scikit-learn   # 可能用于评估或数据处理
requests       # 可能用于下载等网络操作
